pyproject.toml
[tool.poetry]
name = "intentrouter-pro"
version = "0.1.0"
description = "意图驱动的多模态 Agent 编排系统"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "intentrouter"}]

[tool.poetry.dependencies]
python = "^3.11"
# 核心框架
langgraph = "^0.2.0"
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
langchain-community = "^0.3.0"

# Web 框架
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
streamlit = "^1.39.0"

# 数据库
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.35"
pgvector = "^0.3.0"

# 向量与嵌入
faiss-cpu = "^1.9.0"
sentence-transformers = "^3.0.0"

# 多模态
pillow = "^11.0.0"
pypdf = "^5.0.0"

# 工具与实用
pydantic = "^2.9.0"
pydantic-settings = "^2.6.0"
python-dotenv = "^1.0.1"
redis = "^5.0.0"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
black = "^24.10.0"
ruff = "^0.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

